Make unpack_off() handle values > 2GiB. Fixes #2.
This fixes the problem of files larger than 2GiB appearing with negative sizes,
so long as museekd reports the correct size. This accounts for the cases where
other museekd clients were reporting the correct size while museeq was not.

However, some files are still shown with negative sizes in museeq, while the
same files are shown with unbelievably large sizes (0xFFFFFFFF........) in
other museekd clients. Since this seems to happen on a per-peer basis, it seems
that the bug lies either in those peers or in museekd's interpretation of their
messages; I'm not sure which yet. The current official SoulseekQt doesn't show
any files > 2GiB in its search & browse results, so I have no way to be sure
which peers are reporting large file sizes correctly, but a look at the raw
messages from a problematic peer ought to give a strong indication.