museekd: Work around SoulseekQt's filesize=0 bug.
When a peer is ready to send us a file, it first sends us message
code 40: Transfer Request, which includes the file size.  Museekd
and at least one other unofficial Soulseek client (Nicotine) were
using this file size to override the one already on the transfers
list.  I imagine this might have seemed reasonable at some point,
perhaps to handle a remote file that had grown since it was first
placed on the list, but it led to an incompatibility:

The official SoulseekQt client fails to send the correct size for
files larger than 4GiB, sending a zero value instead.  (This is a
bug observed from peers who reported their client software as the
official SoulseekQt versions 2014.8.31 and 12/16/12.  It probably
exists in other versions as well.)

When the zero file size was received, museekd would dutifully use
it as the file size in all future messages, including the message
to initiate the transfer.  The peer (running the official client)
would then choke on that message, causing the transfer to fail.

This patch works around the problem by ignoring filesize=0 in the
Transfer Request message if the file size was previously reported
as nonzero.  We also generate a status message when this happens.